working attach handler
